+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/SentryPlugin.kt
@@ -324,7 +324,7 @@ class SentryPlugin : Plugin<Project> {
     companion object {
         const val SENTRY_ORG_PARAMETER = "sentryOrg"
         const val SENTRY_PROJECT_PARAMETER = "sentryProject"
-        internal const val SENTRY_SDK_VERSION = "6.1.3"
+        internal const val SENTRY_SDK_VERSION = "6.1.4"
 
         internal val sep = File.separator