+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Bump CLI to v2.3.0 ([#338](https://github.com/getsentry/sentry-android-gradle-plugin/pull/338), [#341](https://github.com/getsentry/sentry-android-gradle-plugin/pull/341))
   - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#230)
   - [diff](https://github.com/getsentry/sentry-cli/compare/1.72.0...2.3.0)
+- Bump Android SDK to v6.1.4 ([#340](https://github.com/getsentry/sentry-android-gradle-plugin/pull/340))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#614)
+  - [diff](https://github.com/getsentry/sentry-java/compare/6.1.3...6.1.4)
 
 ### Fixes